backend/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
python-dotenv
psycopg2-binary
marshmallow
gunicorn